requirements.txt
bs4==0.0.2
google-cloud-speech==2.26.0
gtts==2.5.1
mutagen==1.47.0
numpy==1.24.4
opencv-contrib-python==4.9.0.80
openpyxl==3.1.2
pillow==10.3.0
PyAudio==0.2.14
pyautogui==0.9.54
pygame==2.5.2
pywhatkit==5.4
requests==2.31.0
SpeechRecognition==3.10.4
wikipedia==1.4.0